# contracts/projections/openapi.v3.1.yaml
openapi: 3.1.0
info:
  title: Flex Agent representative transport projection
  version: 0.2.0
  description: >
    Deterministic OpenAPI projection of representative command, event, manifest,
    evidence locator, audit, safe error, and SSE shapes. Structural fields mirror
    the canonical JSON Schema catalog; this projection is not authorization evidence
    and excludes internal-only fields by design.
paths:
...
  /v2/assessment/my-work/{enrollmentId}/submission/intake:
    post:
      operationId: beginSubmissionIntake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeginIntakeCommandV2'
      responses:
        '200':
          description: Intake opened or current active intake returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeMutationOutcomeV2'
  /v2/assessment/my-work/{enrollmentId}/submission/intake/{intakeId}/items:
    post:
      operationId: completeSubmissionIntakeItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteIntakeItemCommandV2'
      responses:
        '200':
          description: Item receipt recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeMutationOutcomeV2'
  /v2/assessment/my-work/{enrollmentId}/submission/intake/{intakeId}/cancel:
    post:
      operationId: cancelSubmissionIntake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntakeRevisionCommandV2'
      responses:
        '200':
          description: Intake cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeMutationOutcomeV2'
  /v2/assessment/my-work/{enrollmentId}/submission/intake/{intakeId}/finalize:
    post:
      operationId: finalizeSubmissionIntake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntakeRevisionCommandV2'
      responses:
        '200':
          description: Immutable accepted version committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntakeMutationOutcomeV2'
...
components:
  schemas:
...
    BeginIntakeCommandV2:
      $ref: '../schemas/v2/submission/begin-intake-command.v2.schema.json'
    CompleteIntakeItemCommandV2:
      $ref: '../schemas/v2/submission/complete-intake-item-command.v2.schema.json'
    IntakeRevisionCommandV2:
      $ref: '../schemas/v2/submission/intake-revision-command.v2.schema.json'
    IntakeMutationOutcomeV2:
      $ref: '../schemas/v2/submission/intake-mutation-outcome.v2.schema.json'
...